Prevent flag carriers dropping the flag at (0,0) by spectating
If a flag carrier becomes a spectator, the flag will teleport to (0,0).

SV_FlagDrop calls Unlag::getInstance().getCurrentPlayerPosition() to get
the player's position at which to drop the flag, but the latter returns
early if the player is a spectator, having set the return parameters all
to zero.

Instead have SV_SetPlayerSpec call CTF_CheckFlags *before*
player.spectator is set, so the flag is dropped where the carrier's
feet were when his "finger slipped".

Signed-off-by: RjY <rjy@users.sourceforge.net>